Refactor login process
- move program logic from IndexingForm to IndexingService
- check if index is update when user logs into the system
- check if multiple clients are available after user logs in
