+++ Kitodo/src/main/java/org/kitodo/production/services/ServiceManager.java
@@ -40,6 +40,7 @@ import org.kitodo.production.services.dataeditor.RulesetManagementService;
 import org.kitodo.production.services.dataformat.MetsService;
 import org.kitodo.production.services.file.FileService;
 import org.kitodo.production.services.image.ImageService;
+import org.kitodo.production.services.index.IndexingService;
 import org.kitodo.production.services.migration.MigrationService;
 import org.kitodo.production.services.schema.SchemaService;
 import org.kitodo.production.services.security.SecurityAccessService;
@@ -63,6 +64,7 @@ public class ServiceManager {
     private static FolderService folderService;
     private static ImageService imageService;
     private static ImportService importService;
+    private static IndexingService indexingService;
     private static LdapGroupService ldapGroupService;
     private static LdapServerService ldapServerService;
     private static ListColumnService listColumnService;
@@ -292,6 +294,12 @@ public class ServiceManager {
     private static void initializeCommentService() {
         if (Objects.isNull(commentService)) {
             commentService = CommentService.getInstance();
+        }
+    }
+
+    private static void initializeIndexingService() {
+        if (Objects.isNull(indexingService)) {
+            indexingService = IndexingService.getInstance();
         }
     }
 
@@ -652,5 +660,15 @@ public class ServiceManager {
     public static CommentService getCommentService() {
         initializeCommentService();
         return commentService;
+    }
+
+    /**
+     * Initialize IndexingService if it is not yet initialize and return it.
+     *
+     * @return IndexingService object
+     */
+    public static IndexingService getIndexingService() {
+        initializeIndexingService();
+        return indexingService;
     }
 }
