+++ Kitodo/src/main/resources/messages/messages_de.properties
@@ -690,6 +690,7 @@ renameBatch=Batch umbenennen
 requiredField=Mit * gekennzeichnete Felder sind Pflichtfelder
 resultPDF=Ergebnis-PDF
 resultSet=Ergebnismenge
+retry=Wiederholen
 rights=Rechte
 roman=r\u00F6misch
 role=Rolle
@@ -775,6 +776,10 @@ structureElementAdd=Strukturelement hinzuf\u00FCgen
 structureElementDelete=Strukturelement l\u00F6schen
 strukturelementVerschiebenErlaeuterung=W\u00E4hlen Sie nun die Stelle aus, an die das gew\u00E4hlte Strukturelement verschoben werden soll.
 system=System
+systemErrorHeader=Systemfehler
+systemErrorIndexOutOfDateMessage=Der Suchindex muss aktualisiert werden. Bitte informieren Sie einen Administrator!
+systemMaintenanceHeader=Systemwartung
+systemMaintenanceMessage=Das System wird derzeit gewartet. Bitte versuchen Sie es sp\u00E4ter erneut!
 table=Tabelle
 tableSize=Tabellengr\u00F6\u00DFe
 task=Aufgabe
