+++ Kitodo/src/main/resources/messages/messages_en.properties
@@ -689,6 +689,7 @@ renameBatch=Rename batch
 requiredField=Fields marked with * are required
 resultPDF=Result PDF
 resultSet=Result set
+retry=Retry
 rights=Rights
 roman=Roman
 role=Role
@@ -774,6 +775,10 @@ structureElementAdd=Add docstruct
 structureElementDelete=Delete docstruct
 strukturelementVerschiebenErlaeuterung=Please select the node where the selected docstruct should be pasted.
 system=System
+systemErrorHeader=System error
+systemErrorIndexOutOfDateMessage=Index is out of whack. Please inform an administrator!
+systemMaintenanceHeader=System maintenance
+systemMaintenanceMessage=The system is currently under maintenance. Please try again later!
 table=Table
 tableSize=Table size
 task=Task
