+++ Kitodo/src/main/webapp/WEB-INF/templates/base.xhtml
@@ -72,8 +72,7 @@
                     </ui:insert>
                 </footer>
             </div>
-            <ui:include src="includes/base/baseDeleteDialog.xhtml"/>
-            <ui:include src="includes/base/baseLogoutDialog.xhtml"/>
+            <ui:include src="includes/base/deleteDialog.xhtml"/>
             <ui:include src="/WEB-INF/templates/includes/legal.xhtml"/>
             <ui:insert name="dialog"/>
             <ui:insert name="page-scripts">
