+++ Kitodo/src/main/webapp/WEB-INF/templates/includes/base/deleteDialog.xhtml
