+++ Kitodo/src/main/webapp/WEB-INF/templates/includes/base/indexWarningDialog.xhtml
@@ -1,0 +1,58 @@
+<!--
+ *
+ * (c) Kitodo. Key to digital objects e. V. <contact@kitodo.org>
+ *
+ * This file is part of the Kitodo project.
+ *
+ * It is licensed under GNU General Public License version 3 or later.
+ *
+ * For the full copyright and license information, please read the
+ * GPL3-License.txt file that was distributed with this source code.
+ *
+-->
+
+<ui:composition
+        xmlns="http://www.w3.org/1999/xhtml"
+        xmlns:h="http://xmlns.jcp.org/jsf/html"
+        xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
+        xmlns:p="http://primefaces.org/ui">
+    <p:dialog id="indexWarningDialog"
+              widgetVar="indexWarningDialog"
+              showHeader="false"
+              modal="true"
+              width="500px"
+              appendTo="@(body)"
+              resizable="false"
+              closable="false">
+        <ui:fragment rendered="#{indexingForm.indexingInProgress()}">
+            <h3>#{msgs.systemMaintenanceHeader}</h3>
+            <div class="select-note">#{msgs.systemMaintenanceMessage}</div>
+        </ui:fragment>
+        <ui:fragment rendered="#{not indexingForm.indexingInProgress()}">
+            <h3>#{msgs.systemErrorHeader}</h3>
+            <div class="select-note">#{msgs.systemErrorIndexOutOfDateMessage}</div>
+        </ui:fragment>
+
+        <p:panelGrid>
+            <h:form id="retryForm">
+                <p:panelGrid>
+                    <p:row>
+                        <p:button href="#{request.contextPath}/logout"
+                                  id="logoutButton"
+                                  value="#{msgs.logout}"
+                                  styleClass="primary right"
+                                  style="margin-right: 32px;"
+                                  icon="fa fa-close"
+                                  iconPos="right"/>
+                        <p:commandButton id="retryButton"
+                                         icon="fa fa-repeat fa-lg"
+                                         iconPos="right"
+                                         value="#{msgs.retry}"
+                                         action="#{LoginForm.performPostLoginChecks()}"
+                                         styleClass="secondary right"/>
+                    </p:row>
+                </p:panelGrid>
+            </h:form>
+        </p:panelGrid>
+    </p:dialog>
+</ui:composition>
