+++ Kitodo/src/main/webapp/WEB-INF/templates/includes/header/navigation.xhtml
@@ -136,51 +136,4 @@
             </ul>
         </nav>
     </section>
-    <p:dialog class="client-select-dialog"
-              id="clientSelectDialog"
-              widgetVar="selectClientDialog"
-              showHeader="false"
-              modal="true"
-              width="500px"
-              appendTo="@(body)"
-              resizable="false"
-              closable="false">
-        <h3>#{msgs.welcome}, #{LoginForm.loggedUser.name} #{LoginForm.loggedUser.surname}</h3>
-        <h:form id="select-session-client-form">
-            <div class="select-note">#{msgs.clientSelectNote}</div>
-            <div class="select-selector">
-                <p>#{msgs.clientSelect}</p>
-                <p>
-                    <p:selectOneMenu id="client"
-                                 converter="#{clientConverter}"
-                                 value="#{SessionClientController.selectedClient}"
-                                 autoWidth="false">
-                        <f:selectItems value="#{SessionClientController.availableClientsOfCurrentUser}" var="client" itemValue="#{client}"
-                                       itemLabel="#{client.name}"/>
-                    </p:selectOneMenu>
-                </p>
-                <p:panelGrid>
-                    <p:row>
-                        <p:commandButton id="setSessionClientButton"
-                                         action="#{SessionClientController.setSelectedClientAsSessionClient}"
-                                         onstart="PF('selectClientDialog').hide(); $('#loadingScreen').show();"
-                                         oncomplete="$('#loadingScreen').hide();"
-                                         disabled="#{not SessionClientController.areClientsAvailableForUser()}"
-                                         update="@all"
-                                         styleClass="primary right"
-                                         value="#{msgs.select}"
-                                         icon="fa fa-check"
-                                         iconPos="right"/>
-                        <p:button href="${request.contextPath}/logout"
-                                  id="cancelSessionClientSelectionButton"
-                                  type="button"
-                                  value="#{msgs.cancel}"
-                                  styleClass="secondary right"
-                                  icon="fa fa-close"
-                                  iconPos="right"/>
-                    </p:row>
-                </p:panelGrid>
-            </div>
-        </h:form>
-    </p:dialog>
 </ui:composition>
