+++ Kitodo/src/main/webapp/pages/checks.xhtml
@@ -1,0 +1,34 @@
+<!--
+ *
+ * (c) Kitodo. Key to digital objects e. V. <contact@kitodo.org>
+ *
+ * This file is part of the Kitodo project.
+ *
+ * It is licensed under GNU General Public License version 3 or later.
+ *
+ * For the full copyright and license information, please read the
+ * GPL3-License.txt file that was distributed with this source code.
+ *
+-->
+
+<ui:composition template="/WEB-INF/templates/base.xhtml"
+                xmlns="http://www.w3.org/1999/xhtml"
+                xmlns:f="http://xmlns.jcp.org/jsf/core"
+                xmlns:ui="http://xmlns.jcp.org/jsf/facelets">
+
+    <f:metadata>
+        <f:viewAction action="#{LoginForm.performPostLoginChecks()}"/>
+    </f:metadata>
+
+    <ui:define name="breadcrumbs" />
+
+    <ui:define name="navigation" />
+
+    <ui:define name="content" />
+
+    <ui:define name="dialog">
+        <ui:include src="/WEB-INF/templates/includes/base/indexWarningDialog.xhtml"/>
+        <ui:include src="/WEB-INF/templates/includes/base/selectClientDialog.xhtml"/>
+    </ui:define>
+
+</ui:composition>
