+++ Kitodo/src/main/webapp/pages/desktop.xhtml
@@ -14,14 +14,8 @@
 <ui:composition
         template="/WEB-INF/templates/base.xhtml"
         xmlns:h="http://xmlns.jcp.org/jsf/html"
-        xmlns:f="http://xmlns.jcp.org/jsf/core"
         xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
         xmlns:p="http://primefaces.org/ui">
-
-    <f:metadata>
-        <f:viewAction action="#{indexingForm.checkIndexIntegrity()}"/>
-        <f:viewAction action="#{SessionClientController.showClientSelectDialog()}" />
-    </f:metadata>
 
     <ui:define name="content">
         <h:outputScript name="desktop.js" target="body" library="js"/>
